+++ ios/Classes/CountlyiOS/CountlyConnectionManager.m
@@ -82,6 +82,7 @@ NSString* const kCountlyEndpointWidget = @"/widget";
 NSString* const kCountlyEndpointSurveys = @"/surveys";
 
 const NSInteger kCountlyGETRequestMaxLength = 2048;
+static const NSTimeInterval CONNECTION_TIMEOUT = 30.0;
 
 @implementation CountlyConnectionManager : NSObject
 
@@ -300,6 +301,8 @@ static dispatch_once_t onceToken;
             else
             {
                 CLY_LOG_D(@"%s, request:[ <%p> ] failed! response:[ %@ ]", __FUNCTION__, request, [data cly_stringUTF8]);
+                [CountlyHealthTracker.sharedInstance logFailedNetworkRequestWithStatusCode:((NSHTTPURLResponse*)response).statusCode errorResponse: [data cly_stringUTF8]];
+                [CountlyHealthTracker.sharedInstance saveState];
                 self.startTime = nil;
             }
         }
@@ -317,6 +320,54 @@ static dispatch_once_t onceToken;
 
     [self logRequest:request];
 }
+
+- (BOOL)backoff:(long)responseTimeSeconds queryString:(NSString *)queryString
+{
+    BOOL result = NO;
+    // Check if the current response time is within acceptable limits
+    if (responseTimeSeconds >= [CountlyServerConfig.sharedInstance bomAcceptedTimeoutSeconds]) {
+        // Check if the remaining request count is within acceptable limits
+        NSUInteger remainingRequests = [CountlyPersistency.sharedInstance remainingRequestCount];
+        NSUInteger threshold = (NSUInteger)(CountlyPersistency.sharedInstance.storedRequestsLimit * [CountlyServerConfig.sharedInstance bomRQPercentage]);
+        
+        if (remainingRequests <= threshold) {
+            // Calculate the age of the current request
+            double requestTimestamp = [[queryString cly_valueForQueryStringKey:kCountlyQSKeyTimestamp] longLongValue] / 1000.0;
+            double requestAgeInSeconds = [NSDate date].timeIntervalSince1970 - requestTimestamp;
+            
+            if (requestAgeInSeconds <= [CountlyServerConfig.sharedInstance bomRequestAge] * 3600.0) {
+                // Server is too busy, back off
+                result = YES;
+                [CountlyHealthTracker.sharedInstance logBackoffRequest];
+            }
+        }
+    }
+    
+    if (!result) {
+        [CountlyHealthTracker.sharedInstance logConsecutiveBackoffRequest];
+    }
+    
+    return result;
+}
+
+- (void)backoffCountdown
+{
+    __weak typeof(self) weakSelf = self;
+    CLY_LOG_D(@"%s, backed off, countdown start for %f seconds", __FUNCTION__, [CountlyServerConfig.sharedInstance bomDuration]);
+
+    atomic_store(&_backoff, YES);
+    dispatch_time_t delay = dispatch_time(DISPATCH_TIME_NOW, (int64_t)([CountlyServerConfig.sharedInstance bomDuration] * NSEC_PER_SEC));
+    dispatch_after(delay, dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
+        __strong typeof(weakSelf) strongSelf = weakSelf;
+        if (!strongSelf) return;
+        
+        
+        CLY_LOG_D(@"%s, countdown finished, running tick in background thread", __FUNCTION__);
+        atomic_store(&strongSelf->_backoff, NO);
+        [strongSelf proceedOnQueue];
+    });
+}
+
 
 - (NSString*)extractAndRemoveOverrideEndPoint:(NSString **)queryString
 {