+++ ios/Classes/CountlyiOS/CountlyServerConfig.m
@@ -35,9 +35,15 @@
 @property (nonatomic) NSInteger serverConfigUpdateInterval;
 @property (nonatomic) NSInteger currentServerConfigUpdateInterval;
 
+@property (nonatomic) NSInteger bomAcceptedTimeoutSeconds;
+@property (nonatomic) double bomRQPercentage;
+@property (nonatomic) NSInteger bomRequestAge;
+@property (nonatomic) NSInteger bomDuration;
+
 @property (nonatomic) NSInteger version;
 @property (nonatomic) long long timestamp;
 @property (nonatomic) long long lastFetchTimestamp;
+@property (nonatomic) BOOL serverConfigUpdatesDisabled;
 
 @end
 
@@ -71,6 +77,10 @@ NSString *const kRConsentRequired = @"cr";
 NSString *const kRDropOldRequestTime = @"dort";
 NSString *const kRCrashReporting = @"crt";
 NSString *const kRServerConfigUpdateInterval = @"scui";
+NSString *const kRBOMAcceptedTimeout = @"bom_at";
+NSString *const kRBOMRQPercentage = @"bom_rqp";
+NSString *const kRBOMRequestAge = @"bom_ra";
+NSString *const kRBOMDuration = @"bom_d";
 
 @implementation CountlyServerConfig
 
@@ -92,22 +102,12 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
     self = [super init];
     if (self)
     {
-        // Set default values
-        _trackingEnabled = YES;
-        _networkingEnabled = YES;
-        _crashReportingEnabled = YES;
-        _customEventTrackingEnabled = YES;
-        _enterContentZone = NO;
-        _locationTracking = YES;
-        _viewTrackingEnabled = YES;
-        _sessionTrackingEnabled = YES;
-        _loggingEnabled = NO;
-        _refreshContentZone = YES;
-
         _timestamp = 0;
         _version = 0;
         _currentServerConfigUpdateInterval = 4;
         _requestTimer = nil;
+        _serverConfigUpdatesDisabled = NO;
+        [self setDefaultValues];
     }
     return self;
 }
@@ -144,6 +144,16 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
     {
         *property = value.integerValue;
         [logString appendFormat:@"%@: %ld, ", key, (long)*property];
+    }
+}
+
+- (void)setDoubleProperty:(double *)property fromDictionary:(NSDictionary *)dictionary key:(NSString *)key logString:(NSMutableString *)logString
+{
+    NSNumber *value = dictionary[key];
+    if (value)
+    {
+        *property = value.doubleValue;
+        [logString appendFormat:@"%@: %lf, ", key, (double)*property];
     }
 }
 
@@ -191,6 +201,11 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
     [self setIntegerProperty:&_serverConfigUpdateInterval fromDictionary:dictionary key:kRServerConfigUpdateInterval logString:logString];
     [self setBoolProperty:&_locationTracking fromDictionary:dictionary key:kRLocationTracking logString:logString];
     [self setBoolProperty:&_refreshContentZone fromDictionary:dictionary key:kRRefreshContentZone logString:logString];
+    [self setBoolProperty:&_backoffMechanism fromDictionary:dictionary key:kRbackoffMechanism logString:logString];
+    [self setIntegerProperty:&_bomAcceptedTimeoutSeconds fromDictionary:dictionary key:kRBOMAcceptedTimeout logString:logString];
+    [self setDoubleProperty:&_bomRQPercentage fromDictionary:dictionary key:kRBOMRQPercentage logString:logString];
+    [self setIntegerProperty:&_bomRequestAge fromDictionary:dictionary key:kRBOMRequestAge logString:logString];
+    [self setIntegerProperty:&_bomDuration fromDictionary:dictionary key:kRBOMDuration logString:logString];
 
     CLY_LOG_D(@"%s, version:[%li], timestamp:[%lli], %@", __FUNCTION__, _version, _timestamp, logString);
 }
@@ -297,6 +312,10 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
 
 - (void)fetchServerConfigIfTimeIsUp
 {
+    if (_serverConfigUpdatesDisabled) {
+        return;
+    }
+    
     if (_lastFetchTimestamp)
     {
         long long currentTime = NSDate.date.timeIntervalSince1970 * 1000;
@@ -311,7 +330,13 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
 
 - (void)fetchServerConfig:(CountlyConfig *)config
 {
-    CLY_LOG_D(@"Fetching server configs...");
+    CLY_LOG_D(@"%s, fetching sdk behavior settings", __FUNCTION__);
+    
+    if (_serverConfigUpdatesDisabled) {
+        CLY_LOG_D(@"%s, sdk behavior settings updates disabled, omitting fetch", __FUNCTION__);
+        return;
+    }
+    
     if (CountlyDeviceInfo.sharedInstance.isDeviceIDTemporary)
         return;
 
@@ -349,16 +374,7 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
         if (serverConfigResponse[kRConfig] != nil)
         {
             [CountlyPersistency.sharedInstance storeServerConfig:serverConfigResponse];
-            self->_trackingEnabled = YES;
-            self->_networkingEnabled = YES;
-            self->_crashReportingEnabled = YES;
-            self->_customEventTrackingEnabled = YES;
-            self->_enterContentZone = NO;
-            self->_locationTracking = YES;
-            self->_viewTrackingEnabled = YES;
-            self->_sessionTrackingEnabled = YES;
-            self->_loggingEnabled = NO;
-            self->_refreshContentZone = YES;
+            [self setDefaultValues];
             [self populateServerConfig:serverConfigResponse];
         }
 
@@ -397,6 +413,28 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
     }
 
     CLY_LOG_D(@"serverConfigRequest URL :%@", URL);
+}
+
+- (void)setDefaultValues {
+    _trackingEnabled = YES;
+    _networkingEnabled = YES;
+    _crashReportingEnabled = YES;
+    _customEventTrackingEnabled = YES;
+    _enterContentZone = NO;
+    _locationTracking = YES;
+    _viewTrackingEnabled = YES;
+    _sessionTrackingEnabled = YES;
+    _loggingEnabled = NO;
+    _refreshContentZone = YES;
+    _backoffMechanism = YES;
+    _bomAcceptedTimeoutSeconds = 10;
+    _bomRQPercentage = 0.5;
+    _bomRequestAge = 24;
+    _bomDuration = 60;
+}
+
+- (void)disableSDKBehaviourSettings {
+    _serverConfigUpdatesDisabled = YES;
 }
 
 - (BOOL)trackingEnabled
@@ -506,6 +544,31 @@ NSString *const kRServerConfigUpdateInterval = @"scui";
 - (BOOL)refreshContentZoneEnabled
 {
     return _refreshContentZone;
+}
+
+- (BOOL)backoffMechanism
+{
+    return _backoffMechanism;
+}
+
+- (NSInteger)bomAcceptedTimeoutSeconds
+{
+    return _bomAcceptedTimeoutSeconds;
+}
+
+- (double)bomRQPercentage
+{
+    return _bomRQPercentage;
+}
+
+- (NSInteger)bomRequestAge
+{
+    return _bomRequestAge;
+}
+
+- (NSInteger)bomDuration
+{
+    return _bomDuration;
 }
 
 @end